Moved setting groups into view, added view element, removed all stores
