+++ webviz_subsurface/plugins/_group_tree/_plugin_ids.py
@@ -1,20 +1,5 @@
 class PluginIds:
 
     # pylint: disable=too-few-public-methods
-    class Stores:
-        ENSEMBLES = "ensembles"
-        TREEMODE = "tree-mode"
-        STATISTICS = "statistics"
-        REALIZATION = "realization"
-        FILTER = "filter"
-
-    # pylint: disable=too-few-public-methods
-    class SharedSettings:
-        CONTROLS = "filter"
-        OPTIONS = "options"
-        FILTERS = "filters"
-
-    # pylint: disable=too-few-public-methods
-    class ProductionNetworkID:
-        GROUP_NAME = "Production-Network"
-        GROUP_TREE = "group-tree"
+    class Views:
+        GROUPTREE_VIEW = "group-tree-view"
