--- webviz_subsurface/plugins/_group_tree/shared_settings/__init__.py
@@ -1,3 +1,0 @@
-from ._control import Controls
-from ._filters import Filters
-from ._options import Options
