+++ webviz_subsurface/plugins/_group_tree/view_elements/__init__.py
@@ -1,0 +1,1 @@
+from ._group_tree_view_element import GroupTreeViewElement
