+++ webviz_subsurface/plugins/_group_tree/views/__init__.py
@@ -1,1 +1,1 @@
-from ._group_tree import GroupTreeGraph
+from ._group_tree_view import GroupTreeView
